Token validate operations
When a user logs in, they receive two tokens: an access token and a refresh token. The access token has a 10-minute lifespan, while the refresh token has a one-week lifespan. When the access token expires, a new access token is obtained with a refresh token. Protected endpoints can only be accessed with an access token; refresh tokens cannot be accessed with one. Similarly, only a refresh token is used to obtain a refresh token. This is an application designed to make the system more secure.