Connecting to Redis with Docker and checking the password
Redis was rebuilt with Docker. To prevent brute force attacks, limits were placed on the user's password attempts. When the user entered their password, Redis was accessed to check if the user had any previous incorrect attempts, and if so, how many. If the number of attempts exceeded three, the user was banned for one minute and could continue. If the number of attempts was correct, the number of attempts was reset. The number of attempts was increased by one for each incorrect attempt, and if the number exceeded three, the user was blocked.